Fix logic & gpg behavior

+++ 2021.09-dev/apache/entrypoint-dev.sh
@@ -2,21 +2,21 @@
 set -eu
 
 # just check if we execute apache or php-fpm
-if expr "$1" : "apache" 1>/dev/null || [ "$1" = "php-fpm" ] || [ "${FRIENDICA_DISABLE_UPGRADE:-false}" = "false" ]; then
+if (expr "$1" : "apache" 1>/dev/null || [ "$1" = "php-fpm" ]) && [ "${FRIENDICA_DISABLE_UPGRADE:-false}" = "false" ]; then
   echo "Download sources for ${FRIENDICA_VERSION} (Addon: ${FRIENDICA_ADDONS})"
 
   # Removing the whole directory first
   rm -fr /usr/src/friendica
   export GNUPGHOME="$(mktemp -d)"
 
-  gpg --batch --keyserver keyserver.ubuntu.com --recv-keys 08656443618E6567A39524083EE197EF3F9E4287
+  gpg --batch --logger-fd=1 --no-tty --quiet --keyserver keyserver.ubuntu.com --recv-keys 08656443618E6567A39524083EE197EF3F9E4287
 
   curl -fsSL -o friendica-full-${FRIENDICA_VERSION}.tar.gz "https://files.friendi.ca/friendica-full-${FRIENDICA_VERSION}.tar.gz"
   curl -fsSL -o friendica-full-${FRIENDICA_VERSION}.tar.gz.asc "https://files.friendi.ca/friendica-full-${FRIENDICA_VERSION}.tar.gz.asc";
-  gpg --batch --verify friendica-full-${FRIENDICA_VERSION}.tar.gz.asc friendica-full-${FRIENDICA_VERSION}.tar.gz
+  gpg --batch --logger-fd=1 --no-tty --quiet --verify friendica-full-${FRIENDICA_VERSION}.tar.gz.asc friendica-full-${FRIENDICA_VERSION}.tar.gz
   echo "Core sources (${FRIENDICA_VERSION}) verified"
 
-  tar -xjf friendica-full-${FRIENDICA_VERSION}.tar.gz -C /usr/src/
+  tar -xzf friendica-full-${FRIENDICA_VERSION}.tar.gz -C /usr/src/
   rm friendica-full-${FRIENDICA_VERSION}.tar.gz friendica-full-${FRIENDICA_VERSION}.tar.gz.asc
   mv -f /usr/src/friendica-full-${FRIENDICA_VERSION}/ /usr/src/friendica
   echo "Core sources (${FRIENDICA_VERSION}) extracted"
@@ -25,11 +25,11 @@ if expr "$1" : "apache" 1>/dev/null || [ "$1" = "php-fpm" ] || [ "${FRIENDICA_DI
 
   curl -fsSL -o friendica-addons-${FRIENDICA_ADDONS}.tar.gz "https://files.friendi.ca/friendica-addons-${FRIENDICA_ADDONS}.tar.gz"
   curl -fsSL -o friendica-addons-${FRIENDICA_ADDONS}.tar.gz.asc "https://files.friendi.ca/friendica-addons-${FRIENDICA_ADDONS}.tar.gz.asc"
-  gpg --batch --verify friendica-addons-${FRIENDICA_ADDONS}.tar.gz.asc friendica-addons-${FRIENDICA_ADDONS}.tar.gz
+  gpg --batch --logger-fd=1 --no-tty --quiet --verify friendica-addons-${FRIENDICA_ADDONS}.tar.gz.asc friendica-addons-${FRIENDICA_ADDONS}.tar.gz
   echo "Addon source (${FRIENDICA_ADDONS}) verified"
 
   mkdir -p /usr/src/friendica/addon
-  tar -xjf friendica-addons-${FRIENDICA_ADDONS}.tar.gz -C /usr/src/friendica/addon --strip-components=1
+  tar -xzf friendica-addons-${FRIENDICA_ADDONS}.tar.gz -C /usr/src/friendica/addon --strip-components=1
   rm friendica-addons-${FRIENDICA_ADDONS}.tar.gz friendica-addons-${FRIENDICA_ADDONS}.tar.gz.asc
   echo "Addon sources (${FRIENDICA_ADDONS}) extracted"
